Added unit Test

+++ models/event.go
@@ -15,12 +15,27 @@ type Event struct {
 	UserID      int64     `json:"userId"`
 }
 
+// Hooks for testing
+var (
+	EventSaveFunc          func(e *Event) error
+	EventUpdateFunc        func(e *Event) error
+	EventDeleteFunc        func(e *Event) error
+	GetEventByIDFunc       func(id int64) (*Event, error)
+	GetAllEventsFunc       func() ([]Event, error)
+	CancelRegistrationFunc func(e *Event, userID int64) error
+	RegisterFunc           func(e *Event, userID int64) error
+)
+
 var (
 	GetAllEvents = getAllEvents
 	GetEventByID = getEventByID
 )
 
 func (e *Event) Save() error {
+	if EventSaveFunc != nil {
+		return EventSaveFunc(e)
+	}
+
 	query := `
 	INSERT INTO events(name, description, location, dateTime, user_id) 
 	VALUES (?, ?, ?, ?, ?)`
@@ -43,6 +58,10 @@ func (e *Event) Save() error {
 }
 
 func getAllEvents() ([]Event, error) {
+	if GetAllEventsFunc != nil {
+		return GetAllEventsFunc()
+	}
+
 	query := "SELECT * FROM events"
 	rows, err := db.DB.Query(query)
 	if err != nil {
@@ -67,6 +86,10 @@ func getAllEvents() ([]Event, error) {
 }
 
 func getEventByID(id int64) (*Event, error) {
+	if GetEventByIDFunc != nil {
+		return GetEventByIDFunc(id)
+	}
+
 	query := "SELECT * FROM events WHERE id =?"
 	row := db.DB.QueryRow(query, id)
 
@@ -78,7 +101,11 @@ func getEventByID(id int64) (*Event, error) {
 	return &event, err
 }
 
-func (e Event) Update() error {
+func (e *Event) Update() error {
+	if EventUpdateFunc != nil {
+		return EventUpdateFunc(e)
+	}
+
 	query := `
      Update events  
 SET name = ?, description = ?, location = ?,dateTime =?
@@ -94,7 +121,10 @@ WHERE id = ?
 
 }
 
-func (e Event) Delete() error {
+func (e *Event) Delete() error {
+	if EventDeleteFunc != nil {
+		return EventDeleteFunc(e)
+	}
 
 	query := "DELETE FROM main.events WHERE events.id = ?"
 	Stmt, err := db.DB.Prepare(query)
@@ -107,7 +137,11 @@ func (e Event) Delete() error {
 	_, err = Stmt.Exec(e.ID)
 	return err
 }
-func (e Event) Register(userId int64) error {
+func (e *Event) Register(userId int64) error {
+	if CancelRegistrationFunc != nil {
+		return CancelRegistrationFunc(e, userId)
+	}
+
 	query := `INSERT INTO registration(event_id, user_id) VALUES (?,?)`
 	stmt, err := db.DB.Prepare(query)
 
@@ -122,7 +156,11 @@ func (e Event) Register(userId int64) error {
 	return err
 }
 
-func (e Event) CancelRegistration(userId int64) error {
+func (e *Event) CancelRegistration(userId int64) error {
+	if CancelRegistrationFunc != nil {
+		return CancelRegistrationFunc(e, userId)
+	}
+
 	query := `DELETE FROM registration WHERE event_id = ? AND user_id = ?`
 	stmt, err := db.DB.Prepare(query)
